add filter on pallet module

+++ database/migrations/2024_03_22_014956_create_pallets_table.php
@@ -20,7 +20,7 @@ return new class extends Migration
             $table->date('date');
             $table->integer('allocatedpalletspace');
             $table->integer('spaceuteltotal')->nullable();
-            $table->integer('caseperpallet')->nullable();
+            $table->double('caseperpallet')->nullable();
             $table->timestamps();
         });
     }
